fix: Update Dockerfile to use golang:1.23 and distroless/static
This commit updates the Dockerfile to use golang:1.23 as the base image for the build stage and distroless/static as the base image for the final stage. It also adds the "-o /app/imagepullsecret-patcher" flag to the "go build" command to specify the output binary path.

The changes improve the build process and reduce the final image size.